refactor into functional component
