+++ src/pages/contact.js
@@ -1,13 +1,13 @@
 import React from "react";
 import Layout from "../components/layout";
 import SEO from "../components/seo";
-import ContactForm from "../components/contactForm";
+import Form from "../components/UI/ContactForm/Form";
 
 const Contact = () => {
   return (
     <Layout>
       <SEO title='Contact'></SEO>
-      <ContactForm />
+      <Form />
     </Layout>
   );
 };
